--- qwik-graphql-tailwind/src/routes/[org]/user-page.classNames.ts
@@ -1,4 +1,0 @@
-export const container = 'relative';
-export const stickyNav = 'border-b border-gray-200 sticky top-0 bg-white z-20 hidden md:block';
-export const gridNav = 'grid grid-cols-12 gap-8 max-w-screen-2xl mx-auto';
-export const profileNav = 'col-span-12 md:col-span-8 xl:col-span-9';
