+++ qwik-graphql-tailwind/src/routes/orgs/[name]/index.tsx
@@ -2,6 +2,15 @@ import { component$, useClientEffect$ } from '@builder.io/qwik';
 import { useLocation } from '@builder.io/qwik-city';
 import { OrgProfileCard } from './org-profile-card';
 
+import { ORG_REPOS_QUERY } from '~/utils/queries/org-repos-query';
+import { useQuery } from '~/utils';
+import { GITHUB_GRAPHQL } from '~/utils/constants';
+
+interface OrgRepoQueryParams {
+  organization: string;
+  first: number;
+}
+
 export default component$(() => {
   useClientEffect$(async () => {
     // const abortController = new AbortController();
@@ -15,3 +24,25 @@ export default component$(() => {
     </div>
   );
 });
+
+export async function fetchOrgRepos(
+  { organization, first }: OrgRepoQueryParams,
+  abortController?: AbortController
+): Promise<any> {
+  const { executeQuery$ } = useQuery(ORG_REPOS_QUERY);
+
+  const resp = await executeQuery$({
+    signal: abortController?.signal,
+    url: GITHUB_GRAPHQL,
+    variables: {
+      organization,
+      first,
+    },
+    headersOpt: {
+      Accept: 'application/vnd.github+json',
+      authorization: `Bearer ${sessionStorage.getItem('token')}`,
+    },
+  });
+
+  return await resp.json();
+}
